backend/requirements.txt
Django>=5.0,<6
psycopg2-binary>=2.9
dj-database-url>=2.1
python-dotenv>=1.0
django-environ==0.10.0
gunicorn
Pillow>=10.0.0
django-ckeditor